+++ NEWS.md
@@ -1,6 +1,7 @@
 # GalaxyR 0.1.2
 
 -   Add galaxy_list_files function to list all files from an existing history
+-   Add functionality to delete/purge histories
 
 # GalaxyR 0.1.1